pyproject.toml
[project]
name = "7-capital-compass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
